Allow different exception breakpoints from multiple debuggers to be shown at once
Fixes 126608.

Instead of replacing all exception breakpoints, add new exception breakpoints when a new debug sessions starts.
This way, the existing exception breakpoints stay as is. If no active debug session supports an existing exception breakpoint, it is shown as disabled and can be removed manually.
A currently supported exception breakpoint cannot be removed manually as long as the session is active. This gives more control to the user and saves the breakpoint preferences across different debug sessions in the workspace.